backend/backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==24.2

